+++ .gitignore
@@ -1,3 +1,3 @@
 .vscode/*
 **/.DS_Store
-**/.pem
+**/*.pem
