+++ entity/python/entity_server.py
@@ -18,12 +18,12 @@ from cryptography.hazmat.primitives import padding as pad
 bytes_num = 1024
 rsa_key_size = 256
 
-filesystemManager_dir = {"name" : "", "purpose" : '', "number_key":"", "auth_pubkey_path":"", "privkey_path":"", "auth_ip_address":"", "auth_port_number":"", "port_number":"", "ip_address":"", "network_protocol":""}
+# filesystemManager_dir = {"name" : "", "purpose" : '', "number_key":"", "auth_pubkey_path":"", "privkey_path":"", "auth_ip_address":"", "auth_port_number":"", "port_number":"", "ip_address":"", "network_protocol":""}
 
-distribution_key = {"abs_validity" : "", "cipher_key" : "", "mac_key" : ""}
-session_key = {"sessionkey_id" : "", "abs_validity" : "", "rel_validity" : "", "cipher_key" : "", "mac_key" : ""}
+# distribution_key = {"abs_validity" : "", "cipher_key" : "", "mac_key" : ""}
+# session_key = {"sessionkey_id" : "", "abs_validity" : "", "rel_validity" : "", "cipher_key" : "", "mac_key" : ""}
 
-def load_config(path):
+def load_config(path, filesystemManager_dir):
     f = open(path, 'r')
     while True:
         line = f.readline()
@@ -52,10 +52,9 @@ def load_config(path):
             break
     f.close()
 
-load_config(sys.argv[1])
-print(filesystemManager_dir)
+# load_config(sys.argv[1])
 
-sel = selectors.DefaultSelector()
+# sel = selectors.DefaultSelector()
 
 def write_in_n_bytes(num_key, key_size):
     buffer = bytearray(4)
@@ -92,7 +91,7 @@ def read_unsigned_int_BE(buffer):
        num |= buffer[i] << 8 *(3-i)
     return num
 
-def parse_sessionkey(buffer):
+def parse_sessionkey(buffer, session_key):
     session_key["sessionkey_id"] = buffer[:8] 
     session_key["abs_validity"] = buffer[8:8+6]
     session_key["rel_validity"] = buffer[8+6:8+6+6]
@@ -214,7 +213,7 @@ def parse_sessionkey_id(recv, filesystemManager_dir):
     encrypted_buf = recv[10:]
     return encrypted_buf
 
-def parse_distributionkey(buffer, pubkey, privkey):
+def parse_distributionkey(buffer, pubkey, privkey, distribution_key):
     sign_data = buffer[:rsa_key_size]
     sign_sign = buffer[rsa_key_size:rsa_key_size*2]
     sha256_verify(sign_sign, sign_data, pubkey)
@@ -223,106 +222,106 @@ def parse_distributionkey(buffer, pubkey, privkey):
     distribution_key["cipher_key"] = plaintext[7:7+plaintext[6]]
     distribution_key["mac_key"] = plaintext[8+16:]
 
-def accept_wrapper(sock):
-    conn, addr = sock.accept()
-    print(f"Accepted connection from {addr}")
-    conn.setblocking(False)
-    data = types.SimpleNamespace(addr=addr, inb=b"", outb=b"")
-    events = selectors.EVENT_READ | selectors.EVENT_WRITE
-    sel.register(conn, events, data=data)
+# def accept_wrapper(sock):
+#     conn, addr = sock.accept()
+#     print(f"Accepted connection from {addr}")
+#     conn.setblocking(False)
+#     data = types.SimpleNamespace(addr=addr, inb=b"", outb=b"")
+#     events = selectors.EVENT_READ | selectors.EVENT_WRITE
+#     sel.register(conn, events, data=data)
 
-def service_connection(key, mask):
-    sock = key.fileobj
-    data = key.data
-    global payload_max_num
-    if mask & selectors.EVENT_READ:
-        recv_data = sock.recv(bytes_num)  # Should be ready to read
-        if recv_data:
-            if recv_data[0] == 30:
-                encrypted_buf = parse_sessionkey_id(recv_data, filesystemManager_dir)
-                client_sock = auth_socket_connect(filesystemManager_dir)
-                public_key = load_pubkey(filesystemManager_dir["auth_pubkey_path"])
-                private_key = load_privkey(filesystemManager_dir["privkey_path"])
-                while(1):
-                    recv_data_from_auth = client_sock.recv(1024)
-                    if len(recv_data_from_auth) == 0:
-                        continue
-                    msg_type = recv_data_from_auth[0]
-                    length, length_buf = var_length_int_to_num(recv_data_from_auth[1:])
-                    if msg_type == 0:
-                        nonce_auth = recv_data_from_auth[4+1+length_buf:]
-                        serialize_message, nonce_entity = serialize_message_for_auth(filesystemManager_dir, nonce_auth)
+# def service_connection(key, mask):
+#     sock = key.fileobj
+#     data = key.data
+#     global payload_max_num
+#     if mask & selectors.EVENT_READ:
+#         recv_data = sock.recv(bytes_num)  # Should be ready to read
+#         if recv_data:
+#             if recv_data[0] == 30:
+#                 encrypted_buf = parse_sessionkey_id(recv_data, filesystemManager_dir)
+#                 client_sock = auth_socket_connect(filesystemManager_dir)
+#                 public_key = load_pubkey(filesystemManager_dir["auth_pubkey_path"])
+#                 private_key = load_privkey(filesystemManager_dir["privkey_path"])
+#                 while(1):
+#                     recv_data_from_auth = client_sock.recv(1024)
+#                     if len(recv_data_from_auth) == 0:
+#                         continue
+#                     msg_type = recv_data_from_auth[0]
+#                     length, length_buf = var_length_int_to_num(recv_data_from_auth[1:])
+#                     if msg_type == 0:
+#                         nonce_auth = recv_data_from_auth[4+1+length_buf:]
+#                         serialize_message, nonce_entity = serialize_message_for_auth(filesystemManager_dir, nonce_auth)
                         
-                        ciphertext = asymmetric_encrypt(serialize_message, public_key)
-                        signature = sha256_sign(ciphertext, private_key)
+#                         ciphertext = asymmetric_encrypt(serialize_message, public_key)
+#                         signature = sha256_sign(ciphertext, private_key)
                         
-                        total_buffer = send_sessionkey_request(ciphertext, signature)
-                        client_sock.send(bytes(total_buffer))
+#                         total_buffer = send_sessionkey_request(ciphertext, signature)
+#                         client_sock.send(bytes(total_buffer))
 
-                    elif msg_type == 21:
-                        recv_data = recv_data_from_auth[1+length_buf:]
-                        parse_distributionkey(recv_data, public_key, private_key)
+#                     elif msg_type == 21:
+#                         recv_data = recv_data_from_auth[1+length_buf:]
+#                         parse_distributionkey(recv_data, public_key, private_key)
                         
-                        encrytped_sessionkey = recv_data[rsa_key_size*2:]
+#                         encrytped_sessionkey = recv_data[rsa_key_size*2:]
                         
-                        encrypted_buffer = encrytped_sessionkey[:len(encrytped_sessionkey)-len(distribution_key["mac_key"])]
-                        received_tag = encrytped_sessionkey[len(encrytped_sessionkey)-len(distribution_key["mac_key"]):]
+#                         encrypted_buffer = encrytped_sessionkey[:len(encrytped_sessionkey)-len(distribution_key["mac_key"])]
+#                         received_tag = encrytped_sessionkey[len(encrytped_sessionkey)-len(distribution_key["mac_key"]):]
 
-                        decrypted_buf = symmetric_decrypt_hmac(distribution_key, encrypted_buffer, received_tag)
+#                         decrypted_buf = symmetric_decrypt_hmac(distribution_key, encrypted_buffer, received_tag)
                         
-                        recv_nonce_entity = decrypted_buf[:8]
-                        if nonce_entity != recv_nonce_entity:
-                            print("Failed for communication with Auth")
-                            exit()
-                        else:    
-                            print("Success for communication with Auth")
+#                         recv_nonce_entity = decrypted_buf[:8]
+#                         if nonce_entity != recv_nonce_entity:
+#                             print("Failed for communication with Auth")
+#                             exit()
+#                         else:    
+#                             print("Success for communication with Auth")
 
-                        crypto_buf, crypto_buf_length = var_length_int_to_num(decrypted_buf[8:])
-                        crypto_info = decrypted_buf[8+crypto_buf_length:8+crypto_buf_length+crypto_buf]
-                        print("Crypto Info: ", crypto_info)
-                        sessionkey = decrypted_buf[8+crypto_buf_length+crypto_buf:]
-                        number_of_sessionkey = read_unsigned_int_BE(sessionkey)
-                        print("Number of session key: ", number_of_sessionkey)
-                        parse_sessionkey(sessionkey[4:])
-                        client_sock.close()
+#                         crypto_buf, crypto_buf_length = var_length_int_to_num(decrypted_buf[8:])
+#                         crypto_info = decrypted_buf[8+crypto_buf_length:8+crypto_buf_length+crypto_buf]
+#                         print("Crypto Info: ", crypto_info)
+#                         sessionkey = decrypted_buf[8+crypto_buf_length+crypto_buf:]
+#                         number_of_sessionkey = read_unsigned_int_BE(sessionkey)
+#                         print("Number of session key: ", number_of_sessionkey)
+#                         parse_sessionkey(sessionkey[4:])
+#                         client_sock.close()
 
-                        # first buffer is indicator 1. other buffer is nonce.
-                        dec_buf = symmetric_decrypt_hmac(session_key, encrypted_buf[:32], encrypted_buf[32:])
-                        break
+#                         # first buffer is indicator 1. other buffer is nonce.
+#                         dec_buf = symmetric_decrypt_hmac(session_key, encrypted_buf[:32], encrypted_buf[32:])
+#                         break
                         
-                print("Success for receiving the session key.")
+#                 print("Success for receiving the session key.")
 
-            elif recv_data[0] == 32:
-                data.outb += "Hello"
-                sent = sock.send(data.outb)
-                data.outb = data.outb[sent:]
-        else:
-            print(f"Closing connection to {data.addr}")
-            sel.unregister(sock)
+#             elif recv_data[0] == 32:
+#                 data.outb += "Hello"
+#                 sent = sock.send(data.outb)
+#                 data.outb = data.outb[sent:]
+#         else:
+#             print(f"Closing connection to {data.addr}")
+#             sel.unregister(sock)
 
-host, port = filesystemManager_dir["ip_address"], int(filesystemManager_dir["port_number"])
+# host, port = filesystemManager_dir["ip_address"], int(filesystemManager_dir["port_number"])
 
-lsock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-lsock.bind((host, port))
-lsock.listen()
-print(f"Listening on {(host, port)}")
-lsock.setblocking(False)
-sel.register(lsock, selectors.EVENT_READ, data=None)
+# lsock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+# lsock.bind((host, port))
+# lsock.listen()
+# print(f"Listening on {(host, port)}")
+# lsock.setblocking(False)
+# sel.register(lsock, selectors.EVENT_READ, data=None)
 
-try:
-    while True:
-        events = sel.select(timeout=None)
-        for key, mask in events:
-            if key.data is None:
-                accept_wrapper(key.fileobj)
-            else:
-                service_connection(key, mask)
-except KeyboardInterrupt:
-    print("Caught keyboard interrupt, exiting")
-finally:
-    lsock.close()
-    sel.close()
-    print("Finished")
+# try:
+#     while True:
+#         events = sel.select(timeout=None)
+#         for key, mask in events:
+#             if key.data is None:
+#                 accept_wrapper(key.fileobj)
+#             else:
+#                 service_connection(key, mask)
+# except KeyboardInterrupt:
+#     print("Caught keyboard interrupt, exiting")
+# finally:
+#     lsock.close()
+#     sel.close()
+#     print("Finished")
 
 
 # TODO: Send the message and receive the message -> Success of the secure communication
